requirements.txt
einops
hydra-core >= 1.1
lightning >= 2.0
lightning[extra]
matplotlib >= 3.4
opencv-python >= 4.6
pytorchvideo >= 0.1.5
scikit-learn
seaborn
SoccerNet
tensorboard
timm == 0.6.13
torch >= 2.0
torchaug >= 0.3.1
torchmetrics >= 0.11.0
torchvision >= 0.15
